+++ src/create_tree_view.rs
@@ -1,88 +1,42 @@
 use gtk4::prelude::*;
-use gtk4::TreeViewColumn;
+use gtk4::{CellRendererText, TreeView, TreeViewColumn};
 
 use crate::help_function::{ColumnsResults, ColumnsRules};
 
-pub fn create_tree_view_results(tree_view: &gtk4::TreeView) {
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Type");
-    column.set_resizable(false);
-    column.set_min_width(10);
-    // column.set_sort_column_id(ColumnsResults::Type as i32); // TODO GTK Bug
-    column.add_attribute(&renderer, "text", ColumnsResults::Type as i32);
-    tree_view.append_column(&column);
-
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Current Name");
-    column.set_resizable(true);
-    column.set_min_width(50);
-    // column.set_sort_column_id(ColumnsResults::CurrentName as i32); // TODO GTK BUG
-    column.add_attribute(&renderer, "text", ColumnsResults::CurrentName as i32);
-    tree_view.append_column(&column);
-
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Future Name");
-    column.set_resizable(true);
-    column.set_min_width(50);
-    // column.set_sort_column_id(ColumnsResults::FutureName as i32); // TODO GTK BUG
-    column.add_attribute(&renderer, "text", ColumnsResults::FutureName as i32);
-    tree_view.append_column(&column);
-
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Path");
-    column.set_resizable(true);
-    column.set_min_width(50);
-    // column.set_sort_column_id(ColumnsResults::Path as i32); // TODO GTK BUG
-    column.add_attribute(&renderer, "text", ColumnsResults::Path as i32);
-    tree_view.append_column(&column);
+pub fn create_tree_view_results(tree_view: &TreeView) {
+    // TODO maybe enable sorting, but for now GTK crashes when using it
+    create_default_column(tree_view, ColumnsResults::Type as i32, None, "Type");
+    create_default_column(tree_view, ColumnsResults::CurrentName as i32, None, "Current Name");
+    create_default_column(tree_view, ColumnsResults::FutureName as i32, None, "Future Name");
+    create_default_column(tree_view, ColumnsResults::Path as i32, None, "Path");
 
     tree_view.set_vexpand(true);
 }
 
-pub fn create_tree_view_rules(tree_view: &gtk4::TreeView) {
-    // let renderer = gtk4::CellRendererText::new();
-    // let column: gtk4::TreeViewColumn = TreeViewColumn::new();
-    // column.pack_start(&renderer, true);
-    // column.set_title("Number");
-    // column.set_resizable(true);
-    // column.set_min_width(50);
-    // column.add_attribute(&renderer, "text", ColumnsRules::RuleNumber as i32);
-    // tree_view.append_column(&column);
-
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Tool Type");
-    column.set_resizable(true);
-    column.set_min_width(50);
-    column.add_attribute(&renderer, "text", ColumnsRules::RuleType as i32);
-    tree_view.append_column(&column);
-
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Usage Name");
-    column.set_resizable(true);
-    column.set_min_width(50);
-    column.add_attribute(&renderer, "text", ColumnsRules::UsageType as i32);
-    tree_view.append_column(&column);
-
-    let renderer = gtk4::CellRendererText::new();
-    let column: TreeViewColumn = TreeViewColumn::new();
-    column.pack_start(&renderer, true);
-    column.set_title("Description");
-    column.set_resizable(true);
-    column.set_min_width(50);
-    column.add_attribute(&renderer, "text", ColumnsRules::Description as i32);
-    tree_view.append_column(&column);
+pub fn create_tree_view_rules(tree_view: &TreeView) {
+    create_default_column(tree_view, ColumnsRules::RuleType as i32, None, "Tool Type");
+    create_default_column(tree_view, ColumnsRules::UsageType as i32, None, "Usage Name");
+    create_default_column(tree_view, ColumnsRules::Description as i32, None, "Description");
 
     tree_view.set_vexpand(true);
+}
+
+#[allow(clippy::option_option)]
+fn create_default_column(tree_view: &TreeView, column_id: i32, sort_column_id: Option<Option<i32>>, name: &str) -> (CellRendererText, TreeViewColumn) {
+    let renderer = CellRendererText::new();
+    let column: TreeViewColumn = TreeViewColumn::new();
+    column.pack_start(&renderer, true);
+    column.set_title(name);
+    column.set_resizable(true);
+    column.set_min_width(50);
+    column.add_attribute(&renderer, "text", column_id);
+    if let Some(sort_column_id) = sort_column_id {
+        if let Some(sort_column_id) = sort_column_id {
+            column.set_sort_column_id(sort_column_id);
+        } else {
+            column.set_sort_column_id(column_id);
+        }
+    }
+    tree_view.append_column(&column);
+    (renderer, column)
 }
