+++ src/gui_connection/connect_add_files_button.rs
@@ -7,7 +7,7 @@ use chrono::Local;
 use gtk4::prelude::*;
 use gtk4::ResponseType;
 
-use crate::gui_data::GuiData;
+use crate::gui_data_things::gui_data::GuiData;
 use crate::help_function::{get_list_store_from_tree_view, split_path, ColumnsResults};
 use crate::update_records::{update_records, UpdateMode};
