+++ src/gui_connection/connect_remove_files_button.rs
@@ -1,6 +1,6 @@
 use gtk4::prelude::*;
 
-use crate::gui_data::GuiData;
+use crate::gui_data_things::gui_data::GuiData;
 use crate::help_function::{get_full_file_names_from_selection, remove_selected_rows};
 use crate::update_records::{update_records, UpdateMode};
