+++ src/gui_connection/connect_results_move.rs
@@ -1,6 +1,6 @@
 use gtk4::prelude::*;
 
-use crate::gui_data::GuiData;
+use crate::gui_data_things::gui_data::GuiData;
 use crate::help_function::{get_list_store_from_tree_view, ColumnsResults};
 use crate::update_records::{update_records, UpdateMode};
