+++ src/gui_connection/connect_rule_add.rs
@@ -1,7 +1,8 @@
 use gtk4::prelude::*;
 
-use crate::gui_data::GuiData;
-use crate::help_function::{populate_rules_tree_view, read_rule_from_window};
+use crate::gui_data_things::gui_data::GuiData;
+use crate::help_function::populate_rules_tree_view;
+use crate::rule_read::read_rule_from_window;
 use crate::update_records::{update_records, UpdateMode};
 
 pub fn connect_rule_add(gui_data: &GuiData) {
