+++ src/gui_data_things/class_dialog_rule_purge.rs
