+++ src/gui_data_things/class_dialog_rule_size_letters.rs
