+++ src/gui_data_things/class_dialog_rules.rs
@@ -1,13 +1,13 @@
 use gtk4::{Button, Entry, Label};
 
-use crate::class_dialog_rule_add_number::GuiAddNumber;
-use crate::class_dialog_rule_add_text::GuiAddText;
-use crate::class_dialog_rule_custom::GuiCustom;
-use crate::class_dialog_rule_normalize::GuiNormalize;
-use crate::class_dialog_rule_purge::GuiPurge;
-use crate::class_dialog_rule_replace::GuiReplace;
-use crate::class_dialog_rule_size_letters::GuiSizeLetters;
-use crate::class_dialog_rule_trim::GuiTrim;
+use crate::gui_data_things::class_dialog_rule_add_number::GuiAddNumber;
+use crate::gui_data_things::class_dialog_rule_add_text::GuiAddText;
+use crate::gui_data_things::class_dialog_rule_custom::GuiCustom;
+use crate::gui_data_things::class_dialog_rule_normalize::GuiNormalize;
+use crate::gui_data_things::class_dialog_rule_purge::GuiPurge;
+use crate::gui_data_things::class_dialog_rule_replace::GuiReplace;
+use crate::gui_data_things::class_dialog_rule_size_letters::GuiSizeLetters;
+use crate::gui_data_things::class_dialog_rule_trim::GuiTrim;
 
 #[derive(Clone)]
 pub struct GuiDialogRules {
